refactor(tui): colocate remaining screens' mouse dispatch
Second and final PR of the series colocating mouse dispatch the way
keyboard dispatch already is (issue #310). PR 1 (merged) handled
GistDetail; this one does the remaining six screens: List, Gists,
Pins, Revisions, Help, Config.

Moves apply_navigation_list into screens/list.rs, click_select_list
into the same; and per screen: apply_navigation_<screen>/
click_select_<screen> for gists, pins, revisions, help, config -- each
an impl AppState method in that screen's screens/x.rs, matching PR 1's
detail.rs precedent. keys.rs's apply_navigation/click_select shrink to
one line per arm.

Pins/Gists navigation is a pre-match early-return special case (issue
#274: cannot hold &mut PinsState/GistsManagerState from
`match &mut self.screen` while calling helpers) rather than a match
arm; apply_navigation_pins/apply_navigation_gists reproduce that exact
precompute-len/hmax-then-mutate shape. Revisions/Help/Config kept
their original direct `&mut self.screen` matching (translated to
`let Screen::X(payload) = &mut self.screen else { return false };`);
Pins/Gists kept their original pins_mut()/gist_manager_mut() accessor
style. List needs no guard -- it's a unit variant.

apply_list_cursor_nav (shared by Pins and Gists) and PAGE_SCROLL widen
from private to pub(crate) in keys.rs and stay there, matching Decision
3's precedent that identical cross-screen logic stays put rather than
being force-split. list_page_focused also widens to pub(crate) -- it's
now only ever called from screens/list.rs.

Palette's apply_navigation arm and the Diff/Preview/Confirm shared
scroll arm are untouched, per the issue's explicit exclusions.

No behavior change -- moved bodies are byte-identical, only their
file/visibility changed.

Closes #310
